refactor: FSD 구조에 맞게 API 파일들을 리팩토링
shared/api에는 공통 유틸리티(http.ts)만 존재
