+++ daengdong/features/user/api/useSaveUserInfo.ts
@@ -1,6 +1,6 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { useUserStore } from '@/shared/stores/userStore';
-import { registerUserInfo } from '@/shared/api/user';
+import { registerUserInfo } from '@/entities/user/api/user';
 import { SaveUserParams } from '@/entities/user/model/types';
 
 export const useSaveUserInfo = (isNewUser: boolean) => {
